+++ packages/target/lib/src/raise.dart
@@ -54,3 +54,10 @@ A merge<A>(A Function(Raise<A>) block) => recover(block, id);
 
 Future<A> mergeAsync<A>(Future<A> Function(Raise<A>) block) =>
     recoverAsync(block, id);
+
+extension RaiseEitherExtension<Error, A> on Either<Error, A> {
+  /// Binds this to the given [Raise].
+  ///
+  /// Note: [bind] member function already exists on [Either].
+  A bindTo(Raise<Error> r) => r.bind(this);
+}