+++ packages/target_extension/lib/src/target_either_extensions.dart
@@ -1,4 +1,5 @@
 import 'package:dartz/dartz.dart';
+import 'package:target/target.dart' hide fold;
 import 'package:target_extension/src/target_effects.dart';
 
 extension TargetEitherX<L, R> on Either<L, R> {
@@ -102,5 +103,9 @@ extension TargetNullableTypeToEitherX<R> on R? {
 extension TargetFutureEitherX<L, R> on Future<Either<L, R>> {
   Future<Either<L2, R>> thenLeftMap<L2>(L2 Function(L l) f) {
     return then((it) => it.leftMap(f));
+  }
+
+  Future<R> thenBind(Raise<L> r) {
+    return then(r.bind);
   }
 }